test: reorganize tests related to user task local variables update
Extract tests from `TaskListenerVariablesTest` to dedicated classes.
