+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/bpmn/activity/listeners/task/TaskListenerVariablesTest.java
@@ -10,23 +10,13 @@ package io.camunda.zeebe.engine.processing.bpmn.activity.listeners.task;
 import static io.camunda.zeebe.engine.processing.job.JobCompleteProcessor.TL_JOB_COMPLETION_WITH_VARS_NOT_SUPPORTED_MESSAGE;
 import static java.util.Map.entry;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.groups.Tuple.tuple;
 
 import io.camunda.zeebe.engine.util.EngineRule;
 import io.camunda.zeebe.protocol.Protocol;
-import io.camunda.zeebe.protocol.impl.record.value.usertask.UserTaskRecord;
 import io.camunda.zeebe.protocol.record.Assertions;
-import io.camunda.zeebe.protocol.record.Record;
-import io.camunda.zeebe.protocol.record.RecordType;
 import io.camunda.zeebe.protocol.record.RejectionType;
-import io.camunda.zeebe.protocol.record.ValueType;
 import io.camunda.zeebe.protocol.record.intent.JobIntent;
-import io.camunda.zeebe.protocol.record.intent.UserTaskIntent;
-import io.camunda.zeebe.protocol.record.intent.VariableDocumentIntent;
-import io.camunda.zeebe.protocol.record.intent.VariableIntent;
-import io.camunda.zeebe.test.util.record.RecordingExporter;
 import io.camunda.zeebe.test.util.record.RecordingExporterTestWatcher;
-import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import org.junit.Before;
@@ -249,127 +239,6 @@ public class TaskListenerVariablesTest {
     ENGINE.job().ofInstance(processInstanceKey).withType(listenerType).complete();
   }
 
-  /* VariableDocument update tests */
-  @Test
-  public void shouldRejectVariableUpdateWithPropagateSemanticForUserTask() {
-    // given: a process instance with a Camunda user task
-    final long processInstanceKey =
-        helper.createProcessInstance(helper.createProcessWithZeebeUserTask(t -> t));
-    final var createdUserTaskRecord =
-        RecordingExporter.userTaskRecords(UserTaskIntent.CREATED)
-            .withProcessInstanceKey(processInstanceKey)
-            .getFirst();
-
-    // when: attempting to update variables with 'PROPAGATE' semantic for a user task instance
-    final var variableUpdateRejection =
-        ENGINE
-            .variables()
-            .ofScope(createdUserTaskRecord.getValue().getElementInstanceKey())
-            .withDocument(Map.of("approvalStatus", "SUBMITTED"))
-            .withPropagateSemantic()
-            .expectRejection()
-            .update();
-
-    // then
-    Assertions.assertThat(variableUpdateRejection)
-        .describedAs(
-            "Expect rejection when trying to update variables for a user task instance with 'PROPAGATE' semantic")
-        .hasRecordType(RecordType.COMMAND_REJECTION)
-        .hasValueType(ValueType.VARIABLE_DOCUMENT)
-        .hasRejectionReason(
-            "Expected to update variables for user task with key '%d', but updates with 'PROPAGATE' semantic are not supported yet."
-                .formatted(createdUserTaskRecord.getKey()));
-  }
-
-  @Test
-  public void shouldRejectVariableUpdateWhenUserTaskIsNotInCreatedState() {
-    // given: a process instance with a user task having an assignee and task listeners
-    final long processInstanceKey =
-        helper.createProcessInstance(
-            helper.createProcessWithZeebeUserTask(
-                t ->
-                    t.zeebeAssignee("initial_assignee")
-                        .zeebeTaskListener(l -> l.assigning().type(listenerType + "_assigning"))
-                        .zeebeTaskListener(l -> l.updating().type(listenerType + "_updating"))));
-
-    // since the user task has an initial assignee, the assignment transition is triggered
-    final var assigningUserTaskRecord =
-        RecordingExporter.userTaskRecords(UserTaskIntent.ASSIGNING)
-            .withProcessInstanceKey(processInstanceKey)
-            .getFirst();
-
-    // when: attempting to update variables while the user task is in the 'ASSIGNING' state
-    final var variableUpdateRejection =
-        ENGINE
-            .variables()
-            .ofScope(assigningUserTaskRecord.getValue().getElementInstanceKey())
-            .withDocument(Map.of("employeeId", "E12345"))
-            .withLocalSemantic()
-            .expectRejection()
-            .update();
-
-    // then
-    Assertions.assertThat(variableUpdateRejection)
-        .describedAs(
-            "Expect rejection when trying to update variables for a user task that is currently in the 'ASSIGNING' state")
-        .hasRecordType(RecordType.COMMAND_REJECTION)
-        .hasValueType(ValueType.VARIABLE_DOCUMENT)
-        .hasRejectionReason(
-            "Expected to trigger update transition for user task with key '%d', but it is in state 'ASSIGNING'"
-                .formatted(assigningUserTaskRecord.getKey()));
-  }
-
-  @Test
-  public void
-      shouldUpdateVariablesAndPassUserTaskUpdateTransitionWhenUserTaskHasNoUpdatingListeners() {
-    // given: a process instance with a camunda user task
-    final long processInstanceKey =
-        helper.createProcessInstanceWithVariables(
-            helper.createProcessWithZeebeUserTask(t -> t), Map.of("approvalStatus", "PENDING"));
-
-    final var createdUserTaskRecord =
-        RecordingExporter.userTaskRecords(UserTaskIntent.CREATED)
-            .withProcessInstanceKey(processInstanceKey)
-            .getFirst();
-
-    // when: updating task-scoped variables
-    final var variableUpdateRecord =
-        ENGINE
-            .variables()
-            .ofScope(createdUserTaskRecord.getValue().getElementInstanceKey())
-            .withDocument(Map.of("approvalStatus", "SUBMITTED"))
-            .withLocalSemantic()
-            .update();
-
-    // then: variable update should be successful and trigger the user task update transition
-    Assertions.assertThat(variableUpdateRecord)
-        .describedAs("Expect variables to be successfully updated for a user task")
-        .hasRecordType(RecordType.EVENT)
-        .hasIntent(VariableDocumentIntent.UPDATED)
-        .hasValueType(ValueType.VARIABLE_DOCUMENT);
-
-    Assertions.assertThat(
-            RecordingExporter.variableRecords(VariableIntent.CREATED)
-                .withProcessInstanceKey(processInstanceKey)
-                .withScopeKey(createdUserTaskRecord.getValue().getElementInstanceKey())
-                .getFirst()
-                .getValue())
-        .describedAs("Expect the variable to be created at the local scope of user task element")
-        .hasName("approvalStatus")
-        .hasValue("\"SUBMITTED\"");
-
-    assertThat(
-            RecordingExporter.userTaskRecords()
-                .withProcessInstanceKey(processInstanceKey)
-                .limit(r -> r.getIntent() == UserTaskIntent.UPDATED))
-        .extracting(Record::getIntent, r -> r.getValue().getChangedAttributes())
-        .describedAs(
-            "Expect the user task to pass the update transition with variables as a changed attribute")
-        .containsSequence(
-            tuple(UserTaskIntent.UPDATING, List.of(UserTaskRecord.VARIABLES)),
-            tuple(UserTaskIntent.UPDATED, List.of(UserTaskRecord.VARIABLES)));
-  }
-
   @Test
   public void
       shouldActivateUpdatingListenerJobWithCorrectChangedAttributesHeaderAndVariablesOnTaskVariableUpdate() {
@@ -399,15 +268,12 @@ public class TaskListenerVariablesTest {
           assertThat(job.getCustomHeaders())
               .describedAs("Expect job custom headers to indicate that `variables` were changed")
               .contains(
-                  org.assertj.core.api.Assertions.entry(
-                      Protocol.USER_TASK_CHANGED_ATTRIBUTES_HEADER_NAME, "[\"variables\"]"));
+                  entry(Protocol.USER_TASK_CHANGED_ATTRIBUTES_HEADER_NAME, "[\"variables\"]"));
 
           assertThat(job.getVariables())
               .describedAs(
-                  "Expect job variables to include the updated 'approvalStatus' and retain 'employeeId'")
-              .containsExactly(
-                  org.assertj.core.api.Assertions.entry("approvalStatus", "APPROVED"),
-                  org.assertj.core.api.Assertions.entry("employeeId", "E12345"));
+                  "Expect job variables to include the updated local and retain process variabes")
+              .containsExactly(entry("approvalStatus", "APPROVED"), entry("employeeId", "E12345"));
         });
   }
 }
